+++ migrations/frontend/squashed.sql
@@ -451,66 +451,6 @@ CREATE FUNCTION invalidate_session_for_userid_on_password_change() RETURNS trigg
     END;
 $$;
 
-CREATE FUNCTION iscodyactiveevent(name text) RETURNS boolean
-    LANGUAGE plpgsql IMMUTABLE
-    AS $$
-BEGIN
-  RETURN
-    (name LIKE '%%cody%%' OR name LIKE '%%Cody%%')
-    AND name NOT IN (
-        '%completion:started%',
-        '%completion:suggested%',
-        '%cta%',
-        '%Cta%',
-        'CodyVSCodeExtension:CodySavedLogin:executed',
-        'web:codyChat:tryOnPublicCode',
-        'web:codyEditorWidget:viewed',
-        'web:codyChat:pageViewed',
-        'CodyConfigurationPageViewed',
-        'ClickedOnTryCodySearchCTA',
-        'TryCodyWebOnboardingDisplayed',
-        'AboutGetCodyPopover',
-        'TryCodyWeb',
-        'CodySurveyToastViewed',
-        'SiteAdminCodyPageViewed',
-        'CodyUninstalled',
-        'SpeakToACodyEngineerCTA'
-    );
-END;
-$$;
-
-CREATE FUNCTION iscodyexplanationevent(name text) RETURNS boolean
-    LANGUAGE plpgsql IMMUTABLE
-    AS $$
-BEGIN
-  RETURN name = ANY(ARRAY[
-    'CodyVSCodeExtension:recipe:explain-code-high-level:executed',
-    'CodyVSCodeExtension:recipe:explain-code-detailed:executed',
-    'CodyVSCodeExtension:recipe:find-code-smells:executed',
-    'CodyVSCodeExtension:recipe:git-history:executed',
-    'CodyVSCodeExtension:recipe:rate-code:executed'
-  ]);
-END;
-$$;
-
-CREATE FUNCTION iscodygenerationevent(name text) RETURNS boolean
-    LANGUAGE plpgsql IMMUTABLE
-    AS $$
-BEGIN
-  RETURN name = ANY(ARRAY[
-    'CodyVSCodeExtension:recipe:rewrite-to-functional:executed',
-    'CodyVSCodeExtension:recipe:improve-variable-names:executed',
-    'CodyVSCodeExtension:recipe:replace:executed',
-    'CodyVSCodeExtension:recipe:generate-docstring:executed',
-    'CodyVSCodeExtension:recipe:generate-unit-test:executed',
-    'CodyVSCodeExtension:recipe:rewrite-functional:executed',
-    'CodyVSCodeExtension:recipe:code-refactor:executed',
-    'CodyVSCodeExtension:recipe:fixup:executed',
-	'CodyVSCodeExtension:recipe:translate-to-language:executed'
-  ]);
-END;
-$$;
-
 CREATE FUNCTION merge_audit_log_transitions(internal hstore, arrayhstore hstore[]) RETURNS hstore
     LANGUAGE plpgsql IMMUTABLE
     AS $$
@@ -4305,12 +4245,12 @@ CREATE TABLE repo_embedding_job_stats (
     code_files_embedded integer DEFAULT 0 NOT NULL,
     code_chunks_embedded integer DEFAULT 0 NOT NULL,
     code_files_skipped jsonb DEFAULT '{}'::jsonb NOT NULL,
-    code_bytes_embedded integer DEFAULT 0 NOT NULL,
+    code_bytes_embedded bigint DEFAULT 0 NOT NULL,
     text_files_total integer DEFAULT 0 NOT NULL,
     text_files_embedded integer DEFAULT 0 NOT NULL,
     text_chunks_embedded integer DEFAULT 0 NOT NULL,
     text_files_skipped jsonb DEFAULT '{}'::jsonb NOT NULL,
-    text_bytes_embedded integer DEFAULT 0 NOT NULL,
+    text_bytes_embedded bigint DEFAULT 0 NOT NULL,
     code_chunks_excluded integer DEFAULT 0 NOT NULL,
     text_chunks_excluded integer DEFAULT 0 NOT NULL
 );
@@ -4744,6 +4684,23 @@ CREATE SEQUENCE user_external_accounts_id_seq
     CACHE 1;
 
 ALTER SEQUENCE user_external_accounts_id_seq OWNED BY user_external_accounts.id;
+
+CREATE TABLE user_onboarding_tour (
+    id integer NOT NULL,
+    raw_json text NOT NULL,
+    created_at timestamp without time zone DEFAULT now() NOT NULL,
+    updated_by integer
+);
+
+CREATE SEQUENCE user_onboarding_tour_id_seq
+    AS integer
+    START WITH 1
+    INCREMENT BY 1
+    NO MINVALUE
+    NO MAXVALUE
+    CACHE 1;
+
+ALTER SEQUENCE user_onboarding_tour_id_seq OWNED BY user_onboarding_tour.id;
 
 CREATE TABLE user_pending_permissions (
     id bigint NOT NULL,
@@ -5198,6 +5155,8 @@ ALTER TABLE ONLY temporary_settings ALTER COLUMN id SET DEFAULT nextval('tempora
 ALTER TABLE ONLY user_credentials ALTER COLUMN id SET DEFAULT nextval('user_credentials_id_seq'::regclass);
 
 ALTER TABLE ONLY user_external_accounts ALTER COLUMN id SET DEFAULT nextval('user_external_accounts_id_seq'::regclass);
+
+ALTER TABLE ONLY user_onboarding_tour ALTER COLUMN id SET DEFAULT nextval('user_onboarding_tour_id_seq'::regclass);
 
 ALTER TABLE ONLY user_pending_permissions ALTER COLUMN id SET DEFAULT nextval('user_pending_permissions_id_seq'::regclass);
 
@@ -5721,6 +5680,9 @@ ALTER TABLE ONLY user_emails
 ALTER TABLE ONLY user_external_accounts
     ADD CONSTRAINT user_external_accounts_pkey PRIMARY KEY (id);
 
+ALTER TABLE ONLY user_onboarding_tour
+    ADD CONSTRAINT user_onboarding_tour_pkey PRIMARY KEY (id);
+
 ALTER TABLE ONLY user_pending_permissions
     ADD CONSTRAINT user_pending_permissions_service_perm_object_unique UNIQUE (service_type, service_id, permission, object_type, bind_id);
 
@@ -5925,12 +5887,6 @@ CREATE INDEX discussion_threads_target_repo_repo_id_path_idx ON discussion_threa
 CREATE INDEX event_logs_anonymous_user_id ON event_logs USING btree (anonymous_user_id);
 
 CREATE UNIQUE INDEX event_logs_export_allowlist_event_name_idx ON event_logs_export_allowlist USING btree (event_name);
-
-CREATE INDEX event_logs_name_is_cody_active_event ON event_logs USING btree (iscodyactiveevent(name));
-
-CREATE INDEX event_logs_name_is_cody_explanation_event ON event_logs USING btree (iscodyexplanationevent(name));
-
-CREATE INDEX event_logs_name_is_cody_generation_event ON event_logs USING btree (iscodygenerationevent(name));
 
 CREATE INDEX event_logs_name_timestamp ON event_logs USING btree (name, "timestamp" DESC);
 
@@ -6887,6 +6843,9 @@ ALTER TABLE ONLY user_emails
 
 ALTER TABLE ONLY user_external_accounts
     ADD CONSTRAINT user_external_accounts_user_id_fkey FOREIGN KEY (user_id) REFERENCES users(id);
+
+ALTER TABLE ONLY user_onboarding_tour
+    ADD CONSTRAINT user_onboarding_tour_users_fk FOREIGN KEY (updated_by) REFERENCES users(id);
 
 ALTER TABLE ONLY user_public_repos
     ADD CONSTRAINT user_public_repos_repo_id_fkey FOREIGN KEY (repo_id) REFERENCES repo(id) ON DELETE CASCADE;